+++ src/exceptions/other.py
@@ -20,3 +20,12 @@
 from discord.ext import commands
 
 class InternalCommandError(commands.CommandError): pass
+
+class NotBoundException(InternalCommandError):
+    def __init__(self, obj, msg):
+        super().__init__(f"{obj} not bound: {msg}")
+        self._obj = obj
+
+    @property
+    def obj(self):
+        return self._obj
